Extract override injection into separate script
