+++ scripts/build_generated_scripts.js
@@ -1,8 +1,6 @@
 /* jshint evil:true */
 /* global InspectorBackendClass */
 var fs = require('fs'),
-    execSync = require('child_process').execSync,
-    path = require('path'),
     protocol = require(__dirname + '/../app/devtools/protocol.json');
 
 // Mock objects
@@ -25,35 +23,3 @@ fs.writeFileSync(__dirname + '/../app/devtools/front_end/InspectorBackendCommand
 // Generate empty SupportedCSSProperties.js for now.
 var supportedCSSPropertiesContent = 'WebInspector.CSSMetadata.initializeWithSupportedProperties({});'
 fs.writeFileSync(__dirname + '/../app/devtools/front_end/SupportedCSSProperties.js', supportedCSSPropertiesContent);
-
-// Devtools Overrides
-var appBasePath = path.resolve(__dirname, '..', 'app')
-
-function patchHostModuleDeclaration(){
-  var customFrontEndHostFileName = "NwjsInspectorFrontendHost.js"
-  var frontEndModuleConfigFilePath = appBasePath + '/devtools/front_end/host/module.json'
-  var frontEndModuleConfig = require(frontEndModuleConfigFilePath)
-
-  if(frontEndModuleConfig.scripts.indexOf(customFrontEndHostFileName) === -1){
-    frontEndModuleConfig.scripts.unshift(customFrontEndHostFileName)
-  }
-
-  fs.writeFileSync(
-    frontEndModuleConfigFilePath,
-    JSON.stringify(frontEndModuleConfig, null, 4) + "\n"
-  )
-
-}
-
-function addCustomFrontendHostSymlink(){
-  var overrideFilePath = '../../../devtools_overrides/NwjsInspectorFrontendHost.js'
-  var linkPath = appBasePath + '/devtools/front_end/host/NwjsInspectorFrontendHost.js'
-  var symlinkCommand = "ln -s " +  overrideFilePath + ' ' + linkPath
-
-  if(!fs.existsSync(linkPath)){
-    execSync(symlinkCommand)
-  }
-}
-
-patchHostModuleDeclaration()
-addCustomFrontendHostSymlink()
